+++ src/renderer/src/router/index.ts
@@ -2,12 +2,12 @@ import { RouteRecordRaw, Router, createRouter, createWebHashHistory } from 'vue-
 
 import Home from '@/views/Home.vue'
 import HomePage from '@/views/home/HomePage.vue'
-import Info from '@/views/home/info/MoreInformation.vue'
+import Info from '@/views/home/lib/MoreInformation.vue'
 import Me from '@/views/me/Me.vue'
-import Helper from '@/views/me/helper/Helper.vue'
+import Helper from '@/views/me/lib/Helper.vue'
 import ChooseGoal from '@/views/chooseGoal/ChooseGoal.vue'
-import Template from '@/views/chooseGoal/goalTemplate/Template.vue'
-import TemplateDesc from '@/views/chooseGoal/goalTemplate/goalTemplateDesc/TemplateDesc.vue'
+import Template from '@/views/chooseGoal/lib/Template.vue'
+import TemplateDesc from '@/views/chooseGoal/lib/TemplateDesc.vue'
 import Goals from '@/views/goals/Goals.vue'
 import Dashboard from '@/views/dashboard/Dashboard.vue'
 import Profile from '@/views/dashboard/lib/Profile.vue'
