+++ src/renderer/src/views/chooseGoal/lib/Card.vue
