+++ src/renderer/src/views/chooseGoal/lib/TemplateDesc.vue
