+++ src/renderer/src/views/chooseGoal/lib/template.json
